pyproject.toml
[tool.poetry]
name = "cognix"
version = "0.1.0"
description = "面向企业 AI Agent 的个人工作习惯与偏好记忆引擎"
authors = ["Your Name <your@email.com>"]
readme = "README.md"
packages = [{include = "cognix"}]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
apscheduler = "^3.10.4"
lark-oapi = "^1.2.1"
python-dotenv = "^1.0.1"
langchain = "^0.2.0"
langchain-community = "^0.2.0"

[tool.poetry.scripts]
cognix = "cognix.interfaces.cli.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
